Verify element undisplayed - override timeout

+++ actions/commons/GlobalConstants.java
@@ -16,8 +16,8 @@ public class GlobalConstants {
 	public static final String DB_DEV_USER = "AutomationHQ";
 	public static final String DB_DEV_PASSWORD = "TestP@ssword!";
 
-	public static final long SHORT_TIMEOUT = 5;
-	public static final long LONG_TIMEOUT = 30;
+	public static final int SHORT_TIMEOUT = 5;
+	public static final int LONG_TIMEOUT = 30;
 	public static final long RETRY_TESTFAIL = 3;
 
 }
